[DEV-2269] Refactoring
