+++ src/subdomains/core/buy-crypto/process/services/__tests__/buy-crypto-preparation.service.spec.ts
@@ -7,6 +7,7 @@ import { TestUtil } from 'src/shared/utils/test.util';
 import { BankDataService } from 'src/subdomains/generic/user/models/bank-data/bank-data.service';
 import { FeeService } from 'src/subdomains/supporting/payment/services/fee.service';
 import { TransactionHelper } from 'src/subdomains/supporting/payment/services/transaction-helper';
+import { TransactionRequestService } from 'src/subdomains/supporting/payment/services/transaction-request.service';
 import { PricingService } from 'src/subdomains/supporting/pricing/services/pricing.service';
 import { createCustomBuyCrypto, createDefaultBuyCrypto } from '../../entities/__mocks__/buy-crypto.entity.mock';
 import { BuyCryptoRepository } from '../../repositories/buy-crypto.repository';
@@ -28,6 +29,7 @@ describe('BuyCryptoPreparationService', () => {
   let buyCryptoWebhookService: BuyCryptoWebhookService;
   let feeService: FeeService;
   let buyCryptoService: BuyCryptoService;
+  let transactionRequestService: TransactionRequestService;
 
   /*** Spies ***/
 
@@ -75,6 +77,7 @@ describe('BuyCryptoPreparationService', () => {
     buyCryptoWebhookService = mock<BuyCryptoWebhookService>();
     feeService = mock<FeeService>();
     buyCryptoService = mock<BuyCryptoService>();
+    transactionRequestService = mock<TransactionRequestService>();
 
     const module: TestingModule = await Test.createTestingModule({
       providers: [
@@ -88,6 +91,7 @@ describe('BuyCryptoPreparationService', () => {
         { provide: BuyCryptoWebhookService, useValue: buyCryptoWebhookService },
         { provide: FeeService, useValue: feeService },
         { provide: BuyCryptoService, useValue: buyCryptoService },
+        { provide: TransactionRequestService, useValue: transactionRequestService },
         TestUtil.provideConfig(),
       ],
     }).compile();
