+++ src/subdomains/core/buy-crypto/routes/crypto-route/dto/crypto-payment-info.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
 import { AssetDto } from 'src/shared/models/asset/dto/asset.dto';
 import { MinAmount } from 'src/subdomains/supporting/payment/dto/min-amount.dto';
-import { TransactionError } from 'src/subdomains/supporting/payment/dto/transaction.dto';
+import { TransactionError } from 'src/subdomains/supporting/payment/dto/transaction-error.enum';
 
 export class CryptoPaymentInfoDto {
   @ApiProperty()
