+++ src/subdomains/supporting/payment/dto/transaction-details.dto.ts
@@ -1,5 +1,5 @@
 import { FeeDto } from './fee.dto';
-import { TransactionError } from './transaction.dto';
+import { TransactionError } from './transaction-error.enum';
 
 export interface TargetEstimation {
   sourceAmount: number;
