+++ src/subdomains/supporting/payment/dto/transaction-error.enum.ts
